Subgroup fixes
* fix init_genrand() seed & placement
* fix calculation of tr_idx for subgroup_rotate
* increase the global-work-size to 200000 (2000 not enough to trigger errors)
* clustered_rotate : handle cluster sizes up to 1024

test: revert part of subgroup fixes

+++ test_conformance/subgroups/subgroup_common_templates.h
@@ -524,7 +524,7 @@ template <typename Ty, ShuffleOp operation> struct SHF
                             break;
                         // rotate - treat l as delta
                         case ShuffleOp::rotate:
-                            tr_idx = (i + l) % test_params.subgroup_size;
+                            tr_idx = ((int)i + (int)l) % (int)test_params.subgroup_size;
                             break;
                         case ShuffleOp::clustered_rotate: {
                             tr_idx = ((i & ~(test_params.cluster_size - 1))
